Refaractoring, Updates and documentations
WallhavenAPI struct no more maintaining the response
Started to write the documentations.
In the parameters the size of the seed is set to 7.
Sometimes when the user gives a string having 7 character except null character the compiler will not throw warning
Thought there is a issue with memory safety. So wrote few lines to fix that